Update helper function
